requirements.txt
Flask
wikipedia-api
pandas
beautifulsoup4
nltk
scikit-learn
pyspellchecker
